OR-1108 add UT for basetask and fix bugs
