+++ pkg/tasks/treetask.go
@@ -1,0 +1,9 @@
+package tasks
+
+type NodeTask struct {
+	
+}
+
+type TreeTask struct {
+	Root *BaseTask
+}
